[ML] Fixing reporting of http request errors (#61811)
* [ML] Fixing reporting of http request errors

* fixing types

* updating translations
