+++ x-pack/plugins/ml/public/application/data_frame_analytics/pages/analytics_management/hooks/use_create_analytics_form/index.ts
@@ -4,8 +4,4 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-export {
-  useCreateAnalyticsForm,
-  CreateAnalyticsFormProps,
-  getErrorMessage,
-} from './use_create_analytics_form';
+export { useCreateAnalyticsForm, CreateAnalyticsFormProps } from './use_create_analytics_form';
