+++ x-pack/plugins/ml/public/application/util/ml_error.ts
@@ -7,12 +7,14 @@
 import { KbnError } from '../../../../../../src/plugins/kibana_utils/public';
 
 export class MLRequestFailure extends KbnError {
+  origError: any;
+  resp: any;
   // takes an Error object and and optional response object
   // if error is falsy (null) the response object will be used
   // notify will show the full expandable stack trace of the response if a response object is used and no error is passed in.
-  constructor(error, resp) {
+  constructor(error: any, resp: any) {
     error = error || {};
-    super(error.message || JSON.stringify(resp), MLRequestFailure);
+    super(error.message || JSON.stringify(resp));
 
     this.origError = error;
     this.resp = typeof resp === 'string' ? JSON.parse(resp) : resp;
