+++ x-pack/plugins/translations/translations/ja-JP.json
@@ -7420,7 +7420,6 @@
     "xpack.ml.calendarsEdit.newEventModal.startDateAriaLabel": "開始日",
     "xpack.ml.calendarsEdit.newEventModal.toLabel": "終了:",
     "xpack.ml.calendarsList.deleteCalendars.calendarsLabel": "{calendarsToDeleteCount} 件のカレンダー",
-    "xpack.ml.calendarsList.deleteCalendars.deletingCalendarErrorMessage": "カレンダー {calendarId} の削除中にエラーが発生しました: {errorMessage}",
     "xpack.ml.calendarsList.deleteCalendars.deletingCalendarsNotificationMessage": "{messageId} を削除中",
     "xpack.ml.calendarsList.deleteCalendars.deletingCalendarSuccessNotificationMessage": "{messageId} が選択されました",
     "xpack.ml.calendarsList.deleteCalendarsModal.cancelButtonLabel": "キャンセル",
